Refuse model-wide HVAC name collisions on add/replace; split in-process topology tests
The SDK keeps HVAC component names unique model-wide and silently appends
' 1' on a collision, so a coil named after a component on another loop
came back under a different name than requested. add and replace now check
getModelObjectsByName across all HVACComponents and refuse with the type
that holds the name. replace also built a same-named new component while
the old one still existed, leaving it as 'Clg 1'; it now builds under a
placeholder and takes the final name after the old component is removed.

tests/test_air_loop_supply.py was 536 lines; the in-process topology cases
move to tests/test_air_loop_supply_topology.py (added to CI shard 1).
