+++ .github/workflows/ci.yml
@@ -136,7 +136,7 @@ jobs:
           case ${{ matrix.shard }} in
             1)
               # Mirrors amd64 shard 1: SEB4 sim + component/weather + loop ops + skill_retrofit
-              FILES="tests/test_air_loop_supply.py tests/test_example_workflows.py tests/test_component_properties.py tests/test_comstock.py tests/test_weather.py tests/test_weather_files.py tests/test_mcp_seb4.py tests/test_create_constructions.py tests/test_loop_operations.py tests/test_plant_loop_demand.py tests/test_sizing_properties.py tests/test_skill_retrofit.py tests/test_integration.py"
+              FILES="tests/test_air_loop_supply.py tests/test_air_loop_supply_topology.py tests/test_example_workflows.py tests/test_component_properties.py tests/test_comstock.py tests/test_weather.py tests/test_weather_files.py tests/test_mcp_seb4.py tests/test_create_constructions.py tests/test_loop_operations.py tests/test_plant_loop_demand.py tests/test_sizing_properties.py tests/test_skill_retrofit.py tests/test_integration.py"
               EXTRA_ENV="-e MCP_OSW_PATH=tests/assets/SEB_model/SEB4_baseboard/workflow.osw -e EXPECTED_EUI=1.8750760248144998 -e EXPECTED_EUI_RTOL=0.02 -e EXPECTED_EUI_ATOL=0.0"
               ;;
             2)
@@ -285,7 +285,7 @@ jobs:
               # sim test + component/weather + loop ops + skill_retrofit
               # + load/save model & file listing tools (moved from shard 4 to balance runtime)
               # + merge_coplanar_sliver_surfaces real-fixture regression (moved from shard 4, see timing note above)
-              FILES="tests/test_air_loop_supply.py tests/test_example_workflows.py tests/test_component_properties.py tests/test_comstock.py tests/test_weather.py tests/test_weather_files.py tests/test_mcp_seb4.py tests/test_create_constructions.py tests/test_loop_operations.py tests/test_plant_loop_demand.py tests/test_sizing_properties.py tests/test_skill_retrofit.py tests/test_integration.py tests/test_load_save_model.py tests/test_gbxml_import.py::test_geometry_repair_pipeline_on_austin_apartment_fixture"
+              FILES="tests/test_air_loop_supply.py tests/test_air_loop_supply_topology.py tests/test_example_workflows.py tests/test_component_properties.py tests/test_comstock.py tests/test_weather.py tests/test_weather_files.py tests/test_mcp_seb4.py tests/test_create_constructions.py tests/test_loop_operations.py tests/test_plant_loop_demand.py tests/test_sizing_properties.py tests/test_skill_retrofit.py tests/test_integration.py tests/test_load_save_model.py tests/test_gbxml_import.py::test_geometry_repair_pipeline_on_austin_apartment_fixture"
               EXTRA_ENV="-e MCP_OSW_PATH=tests/assets/SEB_model/SEB4_baseboard/workflow.osw -e EXPECTED_EUI=1.8750760248144998 -e EXPECTED_EUI_RTOL=0.02 -e EXPECTED_EUI_ATOL=0.0"
               ;;
             2)
